Add CopilotSlashCommandProvider for managing runtime slash commands and update related classes
